move remaining useful functions
