+++ tests/geometry/test_deprecated_tools.py
@@ -36,7 +36,6 @@ from bluemira.geometry._deprecated_tools import (
     make_mixed_wire,
     make_wire,
     offset,
-    rotation_matrix,
 )
 from bluemira.geometry.base import BluemiraGeo
 from bluemira.geometry.coordinates import Coordinates, get_area
@@ -65,30 +64,6 @@ class TestArea:
         y = np.array([0, -5, -3, -5, -1, 0, 2, 6, 4, 1])
         with pytest.raises(GeometryError):
             get_area(x, y[:-1])
-
-
-class TestRotationMatrix:
-    def test_axes(self):
-        axes = ["x", "y", "z"]
-        axes2 = [[1, 0, 0], [0, 1, 0], [0, 0, 1]]
-
-        for a1, a2 in zip(axes, axes2):
-            r_1 = rotation_matrix(np.pi / 6, a1)
-            r_2 = rotation_matrix(np.pi / 6, a2)
-            assert np.allclose(r_1, r_2), a1
-
-        axes = ["fail", "somthing", "1"]
-        for axis in axes:
-            with pytest.raises(GeometryError):
-                rotation_matrix(30, axis)
-
-    def test_ccw(self):
-        p1 = [9, 0, 0]
-
-        r_matrix = rotation_matrix(np.pi / 2, axis="z")
-        p2 = r_matrix @ p1
-
-        assert np.isclose(p2[1], 9), p2
 
 
 class TestOffset:
